fix: force search selection on search-form
- This ensures that selecting a search result is required when using that form
