+++ components/manual-form.tsx
@@ -8,162 +8,17 @@ export default function ManualForm({
 }: {
   refetchRequests: () => void;
 }) {
-  const [rememberEmail, setRememberEmail] = useState(false);
-  const { formState, setFormState, status, handleChange, handleSubmit } =
-    useFormHandlers(refetchRequests);
-  //     let timeout: NodeJS.Timeout;
-  //     return (...args: any[]) => {
-  //       clearTimeout(timeout);
-  //       timeout = setTimeout(() => func(...args), delay);
-  //     };
-  //   };
-
-  //   const handleSearch = debounce(async (title: string) => {
-  //     if (!title.trim()) return;
-  //     if (searchQuery.loading) return;
-
-  //     setSearchQuery((prevState) => ({ ...prevState, loading: true }));
-
-  //     try {
-  //       const response = await fetch(`/api/search/${title}`);
-  //       const data = await response.json();
-
-  //       setSearchQuery((prevState) => ({
-  //         ...prevState,
-  //         error: "",
-  //         loading: false,
-  //       }));
-
-  //       setSearchResults(data);
-  //     } catch (err) {
-  //       console.error(err);
-  //       setSearchQuery((prevState) => ({
-  //         ...prevState,
-  //         error: "Failed to search movies",
-  //       }));
-  //     } finally {
-  //       setSearchQuery((prevState) => ({
-  //         ...prevState,
-  //         loading: false,
-  //         error: "",
-  //       }));
-  //       (
-  //         document.getElementById("search_modal") as HTMLDialogElement
-  //       ).showModal();
-  //     }
-  //   }, 300);
-
-  //   const selectResult = (e: React.MouseEvent<HTMLButtonElement>, id: number) => {
-  //     e.preventDefault();
-
-  //     // Get selected media by id
-  //     const selected = searchResults.find((result: any) => result.id === id);
-  //     if (!selected) {
-  //       console.error(`Result with id ${id} not found`);
-  //       return;
-  //     }
-
-  //     const seasonElement = document.getElementById(
-  //       `season-${id}`
-  //     ) as HTMLSelectElement | null;
-
-  //     const getSeasonName = (object: any) => {
-  //       if (seasonElement && seasonElement.value !== "Complete") {
-  //         // Find the selected season data
-  //         const seasonData = object?.seasons.find(
-  //           (season: any) => season.id === parseInt(seasonElement.value)
-  //         );
-
-  //         if (!seasonData) {
-  //           console.error(`Season with id ${seasonElement.value} not found`);
-  //           return;
-  //         }
-  //         return seasonData;
-  //       }
-  //       return "Complete";
-  //     };
-
-  //     const season = getSeasonName(selected);
-
-  //     // Run seperate query if a specific season is selected
-  //     if (season !== "Complete") {
-  //       setFormState((prevState) => ({
-  //         ...prevState,
-  //         title: selected.title + " - " + season.name,
-  //         type: selected.media_type,
-  //         optional: {
-  //           ...(selected.year && {
-  //             year: parseInt(season.air_date.split("-")[0]),
-  //           }),
-  //           ...(selected.poster && { image: season.poster_path }),
-  //           ...(selected.tvcr && { rating: selected.tvcr }),
-  //         },
-  //       }));
-  //     } else {
-  //       // Run standard query if it's a movie or complete series
-  //       setFormState((prevState) => ({
-  //         ...prevState,
-  //         title: selected.title.trim(),
-  //         type: selected.media_type,
-  //         optional: {
-  //           ...(selected.year && { year: parseInt(selected.year) }),
-  //           ...(selected.poster && { image: selected.poster }),
-  //           ...(selected.mpaa && { rating: selected.mpaa }),
-  //           ...(selected.tvcr && { rating: selected.tvcr }),
-  //           ...(selected.seasons && {
-  //             seasons:
-  //               season === "Complete"
-  //                 ? selected.seasons.filter(
-  //                     (season: any) => season.name !== "Specials"
-  //                   )
-  //                 : season,
-  //           }),
-  //         },
-  //       }));
-  //     }
-
-  //     (document.getElementById("search_modal") as HTMLDialogElement).close();
-  //   };
-
-  //   const ratingColor = (rating: number) => {
-  //     switch (true) {
-  //       case rating >= 7:
-  //         return "text-success";
-  //       case rating >= 4 && rating != 7:
-  //         return "text-warning";
-  //       default:
-  //         return "text-error";
-  //     }
-  //   };
-
-  //   const handleSearchChange = (e: any) => {
-  //     const { name, value } = e.target;
-
-  //     setFormState((prevState) => ({
-  //       ...prevState,
-  //       [name]: value,
-  //     }));
-
-  //     handleSearch(value);
-  //   };
-
-  const handleCheckboxChange = (e: any) => {
-    const isChecked = e.target.checked;
-
-    setRememberEmail(isChecked);
-
-    if (isChecked) {
-      localStorage.setItem("email", formState.email);
-    } else {
-      localStorage.removeItem("email");
-    }
-  };
-
-  const updateStoredEmail = () => {
-    if (rememberEmail && formState.email) {
-      localStorage.setItem("email", formState.email);
-    }
-  };
+  const {
+    formState,
+    setFormState,
+    rememberEmail,
+    setRememberEmail,
+    status,
+    handleChange,
+    handleSubmit,
+    handleCheckboxChange,
+    updateStoredEmail,
+  } = useFormHandlers(refetchRequests);
 
   // Check for value from "remember email" checkbox on form
   useEffect(() => {
